// iped-webui/src/app/core/api/generated/model/itemDetails.ts
/**
 * IPED Web API - Phase 1 Initial Contracts
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ItemDetails { 
  [key: string]: any | any;


    itemId: string;
    path?: string;
    mediaType?: string;
    metadata?: { [key: string]: any; };
}