// iped-webui/src/app/core/api/generated/model/renditionLink.ts
/**
 * IPED Web API - Phase 1 Initial Contracts
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RenditionLink { 
    kind: RenditionLinkKindEnum;
    url: string;
}
export enum RenditionLinkKindEnum {
    text = 'text',
    html = 'html',
    image = 'image',
    pdf = 'pdf',
    bytes = 'bytes'
};