// iped-webui/src/app/core/api/generated/model/viewerNavigateHitRequest.ts
/**
 * IPED Web API - Phase 1 Initial Contracts
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ViewerNavigateHitRequest { 
    direction: ViewerNavigateHitRequestDirectionEnum;
    wrap?: boolean;
}
export enum ViewerNavigateHitRequestDirectionEnum {
    next = 'next',
    prev = 'prev'
};