// iped-webui/src/app/core/api/generated/model/viewerOpenResponse.ts
/**
 * IPED Web API - Phase 1 Initial Contracts
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RenditionLink } from './renditionLink';
import { ViewerCapabilities } from './viewerCapabilities';


export interface ViewerOpenResponse { 
    viewerSessionId: string;
    viewerId: string;
    capabilities: ViewerCapabilities;
    renditions?: Array<RenditionLink>;
}